finished initial refactor
