--- simulations/config.ini
@@ -1,40 +1,0 @@
-[INITIAL PARAMETERS]
-; integer number of sources to be simulated
-n_sources = 2e5
-; Minimum simulated flux, in same units as the flux in the observations file ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-fl_min = 5e-5
-; Maximum simulated flux ,in same units as the flux in the observations file  
-fl_max = 5  
-; Fractional error in the simulated flux ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-flux_err = 1e-1  
-; Minimum simulated duration, in same units as the duration in the observations file  
-dmin = 5e-7 
-; Maximum simulated duration, in same units as the duration in the observations file     
-dmax = 5e-2
-; detection threshold, to be multiplied by the noise in the images
-det_threshold = 5 
-; integer, extra detection threshold
-extra_threshold = 0
-; Name to be used for the output files
-file = allint  
-; Must be present as a python file 
-lightcurvetype = tophat
-; only used for choppedgaussian
-gaussiancutoff = 0.1
-; confidence levels for limits
-confidence = 95
-; Upper limit on detections. 3 is 95% confidence
-detections = 5
-
-; The following is only used if no observations are provided in an observations file
-[SIM]
-; Number of observations to be simulated    30 46
-nobs =  46
-; Average sensitivity of simulated observations in Jy 21.7e-6 31.7e-6
-obssens = 10e-6
-; std dev of simulated sens in Jy 4.6e-6 4.6e-6 
-obssig = 2e-6
-; interval between simulated observations in days 7 
-obsinterval = 7
-; duration of observations in days 0.009 0.01417 0.16666666666666666667
-obsdurations =  0.16319444444444445
