--- simulations/empty_observing_file.txt
@@ -1,3 +1,0 @@
-# Observations file.
-# start time (see below)\tduration (sec)\tnoise(Jy)
-# YYYY-MM-DDTHH:MM:SS.SSS\tS\tXX
